+++ tests/testthat/test-surface_hl.R
@@ -1,3 +1,7 @@
+sfc_blues_single <- surface_hl(250)
+sfc_blues_multi <- surface_hl(c(240, 260))
+df_cl <- data.frame(l = c(20, 50, 80), c = c(0, 150, 0))
+
 test_that("surface_hl works()", {
 
   expect_error(surface_hl("foo"), "not a numeric")
@@ -8,27 +12,22 @@ test_that("surface_hl works()", {
   expect_error(surface_hl(250, input_lum = "foo"), "numeric")
   expect_error(surface_hl(250, input_lum = c(20, 20)), "distinct")
 
-  f_single <- surface_hl(250)
-  f_multi <- surface_hl(c(240, 260))
-
-  expect_type(f_single, "closure")
+  expect_type(sfc_blues_single, "closure")
+  expect_s3_class(sfc_blues_single, "cpath_surface_hl")
 
   expect_identical(
-    f_single(c(0, 50, 100)),
+    sfc_blues_single(c(0, 50, 100)),
     c(250, 250, 250)
   )
 
   expect_identical(
-    f_multi(c(0, 50, 100)),
+    sfc_blues_multi(c(0, 50, 100)),
     c(240, 250, 260)
   )
 
 })
 
 test_that("df_hcl() works", {
-
-  sfc_blues_multi <- surface_hl(c(240, 260))
-  df_cl <- data.frame(l = c(20, 50, 80), c = c(0, 150, 0))
 
   expect_error(df_hcl("foo", sfc_blues_multi), "not a data frame")
   expect_error(
@@ -49,3 +48,59 @@ test_that("df_hcl() works", {
   )
 
 })
+
+test_that("mat_chroma() works", {
+
+  hue <- 250
+  luminance <- 50
+  chroma_max <- 86.65
+  step <- 2
+
+  n <- floor(chroma_max / 2)
+
+  mat_cma <- mat_chroma(hue, luminance, chroma_max, step)
+
+  expect_type(mat_cma, "double")
+  expect_true(is.matrix(mat_cma))
+  expect_identical(dimnames(mat_cma), list(NULL, c("h", "c", "l")))
+
+  expect_identical(mat_cma[, "h"], rep(hue, n))
+  expect_identical(mat_cma[, "c"], seq(from = step / 2, by = step, length.out = n))
+  expect_identical(mat_cma[, "l"], rep(luminance, n))
+
+})
+
+test_that("mat_surface_hl() works", {
+
+  expect_error(mat_surface_hl("foo"), "does not inherit")
+
+  mat_sfc <- mat_surface_hl(sfc_blues_multi)
+
+  expect_type(mat_sfc, "double")
+  expect_true(is.matrix(mat_sfc))
+  expect_identical(dimnames(mat_sfc), list(NULL, c("h", "c", "l")))
+
+})
+
+test_that("data_surface_hl() works", {
+
+  expect_error(data_surface_hl("foo"), "does not inherit")
+
+  df_sfc <- data_surface_hl(sfc_blues_multi)
+
+  expect_s3_class(df_sfc, "tbl")
+  expect_named(df_sfc, c("h", "c", "l", "hex"))
+
+})
+
+
+test_that("plot_surface_hl() works", {
+
+  expect_error(plot_surface_hl("foo"), "does not inherit")
+
+  plot_sfc <- plot_surface_hl(sfc_blues_multi)
+
+  expect_s3_class(plot_sfc, "gg")
+
+})
+